pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "surquest-GCP-bq-grid"
version = "0.0.1rc1"
description = "Wrapper around Google Cloud Platform BigQuery client to simplify management of the BigQuery tables."
authors = [
    {name= "Michal Švarc", email= "michal.svarc@surquest.com"}
]
readme = "README.md"
dependencies = [
    "PyYAML >=6.0",
    "google-cloud-bigquery >= 3.5.0",
    "google-cloud-storage >= 2.7.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.2.1",
    "pytest-cov>=4.0.0",
]


[project.urls]
"Homepage" = "https://github.com/surquest/python-gcp-tracer"
"Bug Tracker" = "https://github.com/surquest/python-gcp-tracer/issues"

[tool.hatch.build.targets.wheel.sources]
"src" = ""
"scripts" = "surquest"

[tool.hatch.build.targets.sdist.sources]
"src" = ""
"scripts" = "surquest"
